requirements.txt
# chromadb
streamlit
google-generativeai
langchain
python-dotenv
PyPDF2
